+++ girder/test_girder/test_large_image.py
@@ -606,3 +606,43 @@ def testConfigFileEndpoints(server, admin, fsAssetstore):
         os.environ.pop('GIRDER_CONFIG', None)
         if oldGirderConfig is not None:
             os.environ['GIRDER_CONFIG'] = oldGirderConfig
+
+
+@pytest.mark.usefixtures('unbindLargeImage')
+@pytest.mark.plugin('large_image')
+def testMetadataSearch(server, admin, fsAssetstore):
+    resp = server.request(
+        path='/resource/search', user=admin,
+        params={'q': 'value', 'mode': 'li_metadata', 'types': '["item","folder"]'})
+    assert utilities.respStatus(resp) == 200
+    assert resp.json == {'item': [], 'folder': []}
+    resp = server.request(
+        path='/resource/search', user=admin,
+        params={'q': 'key:key1 value', 'mode': 'li_metadata', 'types': '["item","folder"]'})
+    assert utilities.respStatus(resp) == 200
+    assert resp.json == {'item': [], 'folder': []}
+    resp = server.request(
+        path='/resource/search', user=admin,
+        params={'q': 'key:key2 value', 'mode': 'li_metadata', 'types': '["item","folder"]'})
+    assert utilities.respStatus(resp) == 200
+    assert resp.json == {'item': [], 'folder': []}
+    file = utilities.uploadTestFile('yb10kx5k.png', admin, fsAssetstore)
+    itemId = str(file['itemId'])
+    item = Item().load(id=itemId, force=True)
+    Item().setMetadata(item, {'key1': 'value1'})
+    resp = server.request(
+        path='/resource/search', user=admin,
+        params={'q': 'value', 'mode': 'li_metadata', 'types': '["item","folder"]'})
+    assert utilities.respStatus(resp) == 200
+    assert resp.json != {'item': [], 'folder': []}
+    assert len(resp.json['item']) == 1
+    resp = server.request(
+        path='/resource/search', user=admin,
+        params={'q': 'key:key1 value', 'mode': 'li_metadata', 'types': '["item","folder"]'})
+    assert utilities.respStatus(resp) == 200
+    assert len(resp.json['item']) == 1
+    resp = server.request(
+        path='/resource/search', user=admin,
+        params={'q': 'key:key2 value', 'mode': 'li_metadata', 'types': '["item","folder"]'})
+    assert utilities.respStatus(resp) == 200
+    assert len(resp.json['item']) == 0