+++ girder_annotation/girder_large_image_annotation/__init__.py
@@ -19,7 +19,7 @@ from girder.constants import registerAccessFlag
 from girder.exceptions import ValidationException
 from girder.plugin import GirderPlugin, getPlugin
 from girder.settings import SettingDefault
-from girder.utility import setting_utilities
+from girder.utility import search, setting_utilities
 from girder.utility.model_importer import ModelImporter
 
 from . import constants, handlers
@@ -37,6 +37,15 @@ try:
 except PackageNotFoundError:
     # package is not installed
     pass
+
+
+def metadataSearchHandler(*args, **kwargs):
+    import girder_large_image
+
+    return girder_large_image.metadataSearchHandler(
+        models=['item'],
+        searchModels={('annotation', 'large_image'): {'model': 'item', 'reference': 'itemId'}},
+        metakey='annotation.attributes', *args, **kwargs)
 
 
 # Validators
@@ -93,3 +102,6 @@ class LargeImageAnnotationPlugin(GirderPlugin):
         events.bind(
             'data.process', 'large_image_annotation.annotations',
             handlers.process_annotations)
+
+        search._allowedSearchMode.pop('li_annotation_metadata', None)
+        search.addSearchMode('li_annotation_metadata', metadataSearchHandler)